order view, order_card, Fixes some bugs, improves the stability of software operation, and improves poor UI design

+++ lib/model/goods_detail.dart
@@ -115,9 +115,9 @@ class GoodsInfo {
   int id;
   String name;
   String info;
-  int price;
+  double price;
   String ownerId;
-  int originalPrice;
+  double originalPrice;
   String defaultImage;
   int count;
   int newPercentage;
